if true then run build pack

+++ cloudezz-tomcat7/init.sh
@@ -1,8 +1,13 @@
 #!/bin/sh
 
+# start Shell in a box service only when the webshell env is set to true
+if [ $RUN_BUILD_PACK == "true" ]
+then
 rm -r /opt/build-pack >/dev/null 2>/dev/null
 wget -P /opt/build-pack/ https://github.com/cloudezz/cloudezz-tomcat7-build-pack/archive/master.zip  >/dev/null 2>/dev/null
 chmod +x -R /opt/build-pack/
 unzip /opt/build-pack/master.zip -d /opt/build-pack/  >/dev/null 2>/dev/null
 chmod +x -R /opt/build-pack/ 
 cd /opt/build-pack/cloudezz-tomcat7-build-pack-master && ant 
+fi
+
